+++ src/main/java/org/betonquest/betonquest/item/typehandler/FlagHandler.java
@@ -4,9 +4,9 @@ import org.betonquest.betonquest.exceptions.InstructionParseException;
 import org.betonquest.betonquest.item.QuestItem;
 import org.bukkit.inventory.ItemFlag;
 import org.bukkit.inventory.meta.ItemMeta;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.Arrays;
-import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -22,12 +22,13 @@ public class FlagHandler {
     /**
      * Existence of the flags.
      */
-    private final QuestItem.Existence existence = QuestItem.Existence.WHATEVER;
+    private QuestItem.Existence existence = QuestItem.Existence.WHATEVER;
 
     /**
      * Construct a new FlagHandler.
      */
     public FlagHandler() {
+        itemFlags = Set.of();
     }
 
     /**
@@ -37,7 +38,7 @@ public class FlagHandler {
      * @throws InstructionParseException If there is an error parsing.
      */
     public void parse(final String data) throws InstructionParseException {
-        this.itemFlags = Arrays.stream(data.split(",")).map(ItemFlag::valueOf).collect(Collectors.toSet());
+        set(Arrays.stream(data.split(",")).map(ItemFlag::valueOf).collect(Collectors.toSet()));
     }
 
     /**
@@ -46,11 +47,13 @@ public class FlagHandler {
      * @param itemFlags The ItemFlags, or null if not set.
      * @throws InstructionParseException If there is an error setting the flags.
      */
-    public void set(final Set<ItemFlag> itemFlags) throws InstructionParseException {
+    public void set(@Nullable final Set<ItemFlag> itemFlags) throws InstructionParseException {
         if (itemFlags == null || itemFlags.isEmpty()) {
             this.itemFlags = Set.of();
+            this.existence = QuestItem.Existence.FORBIDDEN;
         } else {
-            this.itemFlags = new HashSet<>(itemFlags);
+            this.itemFlags = Set.copyOf(itemFlags);
+            this.existence = QuestItem.Existence.REQUIRED;
         }
     }
 
@@ -72,7 +75,7 @@ public class FlagHandler {
     public boolean check(final ItemMeta data) {
         return existence == QuestItem.Existence.WHATEVER
                 || existence == QuestItem.Existence.FORBIDDEN && data.getItemFlags().isEmpty()
-                || existence == QuestItem.Existence.REQUIRED && (data.getItemFlags().size() > 0) && itemFlags.equals(data.getItemFlags());
+                || existence == QuestItem.Existence.REQUIRED && !data.getItemFlags().isEmpty() && itemFlags.equals(data.getItemFlags());
     }
 
 }